Refactor various utils and types into dedicated files
